refactor: make search engine sources crate-independent
