+++ .github/workflows/unit-tests.yml
@@ -34,7 +34,7 @@ jobs:
       - id: check
         run: |
           CHANGED=$(git diff --name-only origin/${{ github.base_ref || 'main' }} ${{ github.sha }})
-          if [ -z "$CHANGED" ] || echo "$CHANGED" | grep -qE '^src/|^proto/|^\.cargo/|^Cargo\.(toml|lock)$|^build\.rs$|^rust-toolchain\.toml$'; then
+          if [ -z "$CHANGED" ] || echo "$CHANGED" | grep -qE '^src/|^proto/|^\.cargo/|^Cargo\.(toml|lock)$|^build\.rs$|^rust-toolchain\.toml$|^scripts/check-search-engine-boundary\.sh$'; then
             echo "has_backend_changes=true" >> $GITHUB_OUTPUT
           else
             echo "has_backend_changes=false" >> $GITHUB_OUTPUT
@@ -97,6 +97,9 @@ jobs:
 
       - name: Check code formatting
         run: cargo fmt --all -- --check
+
+      - name: Check search engine boundary
+        run: scripts/check-search-engine-boundary.sh
 
       - name: Run clippy
         run: cargo --config net.git-fetch-with-cli=true clippy -- -D warnings
