+++ src/api_query/src/search/mod.rs
@@ -2027,7 +2027,10 @@ pub async fn result_schema(
             }
         };
 
-    let res_schema = match get_result_schema(sql, is_streaming, use_cache).await {
+    let query_context = crate::service::search::datafusion::context::QueryExecutionContext::new(
+        crate::service::search::prepare_query_transforms(&org_id),
+    );
+    let res_schema = match get_result_schema(sql, is_streaming, use_cache, &query_context).await {
         Ok(v) => v,
         Err(e) => {
             return (
