+++ src/core/Cargo.toml
@@ -20,12 +20,13 @@ enterprise = [
     "search/enterprise",
 ]
 cloud = ["infra/cloud", "enterprise", "o2_enterprise/cloud"]
-vectorscan = ["o2_enterprise/vectorscan"]
+vectorscan = ["o2_enterprise/vectorscan", "search/vectorscan"]
 vortex = [
     "dep:vortex",
     "dep:vortex-datafusion",
     "config/vortex",
     "o2_enterprise?/vortex",
+    "search/vortex",
 ]
 
 [dependencies]
