+++ src/core/src/service/search/datafusion/exec.rs
@@ -56,8 +56,10 @@ use {
 };
 
 use super::{
-    peak_memory_pool::PeakMemoryPool, planner::extension_planner::OpenobserveQueryPlanner,
-    storage::file_list, udf::transform_udf::get_all_transform,
+    peak_memory_pool::PeakMemoryPool,
+    planner::extension_planner::OpenobserveQueryPlanner,
+    storage::file_list,
+    udf::transform_udf::{PreparedQueryTransform, get_all_transform},
 };
 use crate::service::search::{
     datafusion::{
@@ -286,7 +288,7 @@ impl<'a> DataFusionContextBuilder<'a> {
     }
 }
 
-pub fn register_udf(ctx: &SessionContext, org_id: &str) -> Result<()> {
+pub fn register_udf(ctx: &SessionContext, transforms: Vec<PreparedQueryTransform>) -> Result<()> {
     ctx.register_udf(super::udf::str_match_udf::STR_MATCH_UDF.clone());
     ctx.register_udf(super::udf::str_match_udf::STR_MATCH_IGNORE_CASE_UDF.clone());
     ctx.register_udf(super::udf::fuzzy_match_udf::FUZZY_MATCH_UDF.clone());
@@ -314,7 +316,7 @@ pub fn register_udf(ctx: &SessionContext, org_id: &str) -> Result<()> {
         super::udaf::summary_percentile::SummaryPercentile::new(),
     ));
     ctx.register_udf(super::udf::cast_to_timestamp_udf::CAST_TO_TIMESTAMP_UDF.clone());
-    let udf_list = get_all_transform(org_id)?;
+    let udf_list = get_all_transform(transforms)?;
     for udf in udf_list {
         ctx.register_udf(udf.clone());
     }
@@ -768,7 +770,7 @@ mod tests {
     #[tokio::test]
     async fn test_register_udf() -> Result<()> {
         let ctx = SessionContext::new();
-        let result = register_udf(&ctx, "test_org");
+        let result = register_udf(&ctx, Vec::new());
 
         assert!(result.is_ok());
