+++ src/core/src/service/search/grpc/flight.rs
@@ -72,7 +72,7 @@ use crate::service::{
         grpc::QueryParams,
         index::IndexCondition,
         inspector::{SearchInspectorFieldsBuilder, search_inspector_fields},
-        match_file,
+        match_file, prepare_query_transforms,
     },
 };
 
@@ -98,7 +98,7 @@ pub async fn search(
         .await?;
 
     // register udf
-    register_udf(&ctx, &org_id)?;
+    register_udf(&ctx, prepare_query_transforms(&org_id))?;
     datafusion_functions_json::register_all(&mut ctx)?;
 
     // Decode physical plan from bytes
