+++ src/core/src/service/search/grpc/mod.rs
@@ -16,14 +16,8 @@
 use std::sync::Arc;
 
 use arrow_schema::Schema;
-use config::meta::{
-    search::ScanStats,
-    stream::{FileKey, StreamType},
-};
-use datafusion::{
-    datasource::TableProvider, execution::cache::cache_manager::FileStatisticsCache,
-    sql::TableReference,
-};
+use config::meta::{search::ScanStats, stream::FileKey};
+use datafusion::{datasource::TableProvider, execution::cache::cache_manager::FileStatisticsCache};
 use hashbrown::HashSet;
 use infra::errors::Result;
 
@@ -32,28 +26,11 @@ use super::{datafusion::exec::TableBuilder, index::IndexCondition};
 pub mod flight;
 pub mod storage;
 pub mod wal;
+pub use ::search::{file_cache::calc_target_partitions, types::QueryParams};
+
 pub use super::tantivy::cache as tantivy_result_cache;
 
 pub type SearchTable = Result<(Vec<Arc<dyn TableProvider>>, ScanStats, HashSet<u64>)>;
-
-#[derive(Debug)]
-pub struct QueryParams {
-    pub trace_id: String,
-    pub org_id: String,
-    pub stream: TableReference,
-    pub stream_type: StreamType,
-    pub stream_name: String,
-    pub time_range: (i64, i64),
-    pub work_group: Option<String>,
-    pub use_inverted_index: bool,
-}
-
-/// Linear interpolation: cached_ratio=0 -> query_thread_num, cached_ratio=1 -> cpu_num.
-pub fn calc_target_partitions(cpu_num: usize, query_thread_num: usize, cached_ratio: f64) -> usize {
-    (cpu_num as i64
-        + ((query_thread_num as i64 - cpu_num as i64) as f64 * (1.0 - cached_ratio)) as i64)
-        as usize
-}
 
 /// Create tables from files, automatically splitting them based on time range overlap:
 /// - Files completely within the query time range: no timestamp filter applied
