+++ src/core/src/service/search/grpc/storage.rs
@@ -15,10 +15,11 @@
 
 use std::sync::Arc;
 
+pub use ::search::file_cache::{cache_files, calc_target_partitions};
 use arrow_schema::Schema;
 use config::{
     cluster::LOCAL_NODE,
-    get_config, is_local_disk_storage,
+    get_config,
     meta::{
         inverted_index::IndexOptimizeMode,
         search::{ScanStats, StorageType},
@@ -36,7 +37,6 @@ use infra::{
 use itertools::Itertools;
 use tracing::Instrument;
 
-pub use super::calc_target_partitions;
 pub use crate::service::search::tantivy::tantivy_search;
 use crate::service::{
     file_list,
@@ -295,134 +295,6 @@ pub async fn search(
         )
     );
     Ok((tables, scan_stats, HashSet::new()))
-}
-
-#[tracing::instrument(name = "service:search:grpc:storage:cache_files", skip_all)]
-pub async fn cache_files(
-    trace_id: &str,
-    files: &[(i64, &String, &String, i64, i64)],
-    scan_stats: &mut ScanStats,
-    file_type: &str,
-) -> (file_data::CacheType, u64, u64) {
-    // check how many files already cached
-    let mut cached_files = HashSet::with_capacity(files.len());
-    let (mut cache_hits, mut cache_misses) = (0, 0);
-
-    let start = std::time::Instant::now();
-    for (_id, _account, file, _size, max_ts) in files.iter() {
-        if file_data::memory::exist(file).await {
-            scan_stats.querier_memory_cached_files += 1;
-            cached_files.insert(file);
-            cache_hits += 1;
-        } else if file_data::disk::exist(file).await {
-            scan_stats.querier_disk_cached_files += 1;
-            cached_files.insert(file);
-            cache_hits += 1;
-        } else {
-            cache_misses += 1;
-        };
-
-        // Record file access metrics
-        let stream_type = if file_type == "index" {
-            config::meta::stream::StreamType::Index
-        } else {
-            // Determine stream type from the file path
-            if file.contains("/logs/") {
-                config::meta::stream::StreamType::Logs
-            } else if file.contains("/metrics/") {
-                config::meta::stream::StreamType::Metrics
-            } else if file.contains("/traces/") {
-                config::meta::stream::StreamType::Traces
-            } else {
-                config::meta::stream::StreamType::Logs // Default
-            }
-        };
-
-        let current_time = chrono::Utc::now().timestamp_micros();
-        let file_age_seconds = (current_time - max_ts) / 1_000_000;
-        let file_age_hours = file_age_seconds as f64 / 3600.0;
-
-        if file_age_hours > 0.0 {
-            config::metrics::FILE_ACCESS_TIME
-                .with_label_values(&[&stream_type.to_string()])
-                .observe(file_age_hours);
-        }
-    }
-
-    let check_cache_took = start.elapsed().as_millis() as usize;
-    if check_cache_took > 1000 {
-        log::warn!(
-            "[trace_id {trace_id}] search->storage: check file cache took: {check_cache_took} ms",
-        );
-    }
-
-    let files_num = files.len() as i64;
-    if files_num == scan_stats.querier_memory_cached_files + scan_stats.querier_disk_cached_files {
-        // all files are cached
-        return (file_data::CacheType::Disk, cache_hits, cache_misses);
-    }
-
-    // check cache size
-    let cfg = get_config();
-    let cache_type = if cfg.memory_cache.enabled
-        && scan_stats.compressed_size < cfg.memory_cache.skip_size as i64
-    {
-        // if scan_compressed_size < ZO_MEMORY_CACHE_SKIP_SIZE, use memory cache
-        file_data::CacheType::Memory
-    } else if !is_local_disk_storage()
-        && cfg.disk_cache.enabled
-        && scan_stats.compressed_size < cfg.disk_cache.skip_size as i64
-    {
-        // if scan_compressed_size < ZO_DISK_CACHE_SKIP_SIZE, use disk cache
-        file_data::CacheType::Disk
-    } else {
-        // no cache, the files are too big than cache size
-        return (file_data::CacheType::None, cache_hits, cache_misses);
-    };
-
-    let trace_id = trace_id.to_string();
-    let files = files
-        .iter()
-        .filter_map(|(id, account, file, size, ts)| {
-            if cached_files.contains(&file) {
-                None
-            } else {
-                Some((*id, account.to_string(), file.to_string(), *size, *ts))
-            }
-        })
-        .collect_vec();
-    let file_type = file_type.to_string();
-    tokio::spawn(async move {
-        let files_num = files.len();
-        for (id, account, file, size, ts) in files {
-            if let Err(e) = crate::service::file_downloader::queue_download(
-                trace_id.clone(),
-                id,
-                account,
-                file.clone(),
-                size,
-                ts,
-                cache_type,
-            )
-            .await
-            {
-                log::error!(
-                    "[trace_id {trace_id}] error in queuing file {file} for background download: {e}"
-                );
-            }
-        }
-        log::info!(
-            "[trace_id {trace_id}] search->storage: successfully enqueued {files_num} files of {file_type} for background download into {cache_type:?}",
-        );
-    });
-
-    // if cached file less than 50% of the total files, return None
-    if scan_stats.querier_memory_cached_files + scan_stats.querier_disk_cached_files < files_num / 2
-    {
-        (file_data::CacheType::None, cache_hits, cache_misses)
-    } else {
-        (cache_type, cache_hits, cache_misses)
-    }
 }
 
 /// Check bloom filter for the file list
