+++ src/search/src/cipher/registry.rs
