+++ src/search/src/lib.rs
@@ -13,7 +13,11 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+#[cfg(feature = "enterprise")]
+pub mod cipher;
 pub mod enrichment;
+pub mod file_cache;
+pub mod types;
 
 use config::meta::{
     search::{Response, ScanStats},
