+++ src/search/src/types.rs
@@ -5,7 +5,7 @@
 // the Free Software Foundation, either version 3 of the License, or
 // (at your option) any later version.
 //
-// This program is distributed in the hope that it will be useful
+// This program is distributed in the hope that it will be useful,
 // but WITHOUT ANY WARRANTY; without even the implied warranty of
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU Affero General Public License for more details.
@@ -13,14 +13,17 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-pub mod bytes_cache;
-pub mod file_data;
-pub mod file_downloader;
-pub mod meta;
-pub mod stats;
-pub mod storage;
+use config::meta::stream::StreamType;
+use datafusion::sql::TableReference;
 
-pub async fn init() -> Result<(), anyhow::Error> {
-    file_data::init().await?;
-    Ok(())
+#[derive(Debug)]
+pub struct QueryParams {
+    pub trace_id: String,
+    pub org_id: String,
+    pub stream: TableReference,
+    pub stream_type: StreamType,
+    pub stream_name: String,
+    pub time_range: (i64, i64),
+    pub work_group: Option<String>,
+    pub use_inverted_index: bool,
 }
